feat: seprate input select field to widget
